added day forecast

+++ core_network/src/main/java/ru/spectr/core_network/models/ConsolidatedWeather.kt
@@ -6,16 +6,16 @@ data class ConsolidatedWeather(
     val applicable_date: String?,
     val created: String?,
     val humidity: Int?,
-    val max_temp: Double?,
-    val min_temp: Double?,
+    val max_temp: Float?,
+    val min_temp: Float?,
     val predictability: Int?,
     val the_temp: Float?,
-    val visibility: Double?,
+    val visibility: Float?,
     val weather_state_abbr: String?,
     val weather_state_name: String?,
-    val wind_direction: Double?,
+    val wind_direction: Float?,
     val wind_direction_compass: String?,
-    val wind_speed: Double?
+    val wind_speed: Float?
 ) {
     companion object {
         const val DATE_PATTERN = "yyyy-MM-dd"
